+++ src/gfx/OceanPipeline.cpp
@@ -69,7 +69,7 @@ void gfx::OceanPipeline::setGeometry(const std::vector<OceanVertex> &verts, cons
         indices.data(), indices.size() * sizeof(uint32_t),
         VK_BUFFER_USAGE_INDEX_BUFFER_BIT,
         m_index_buffer, m_index_buffer_memory);
-    m_num_indices = indices.size();
+    m_num_indices = static_cast<uint32_t>(indices.size());
 }
 
 void gfx::OceanPipeline::setTransform(const glm::mat4x4 &xform) {
